refactor: modify collection page (former main) and add a new main page
the old main page is now the collection page, and the new main page is a hub for all flash cards collections, also, the root is no longer the login page, now it is just a server-side code that redirect to the login, with credentials verification to be added. furthermore, I added a path configuration in the tsconfig.json to use absolute path, this was needed because I also moved the type definition at collection/[id] to an external ts file. Todo: polish this empty front-end and finally start the back-end development